Complete A-Packages
* duke for Duke
* duke.exception for all exception classes
* duke.task for all task related classes
* duke.util for Ui, Parser and Storage
