+++ src/main/java/duke/exception/DataIntegrityException.java
@@ -1,0 +1,7 @@
+package duke.exception;
+
+public class DataIntegrityException extends DukeException {
+    public DataIntegrityException() {
+        super("Storage file integrity compromised :(");
+    }
+}
