+++ src/main/java/duke/exception/IllegalFormatException.java
@@ -1,0 +1,7 @@
+package duke.exception;
+
+public class IllegalFormatException extends DukeException {
+    public IllegalFormatException(String format) {
+        super("Please follow this format:\n  " + format);
+    }
+}
