+++ src/main/java/duke/util/CheckedFunction.java
@@ -1,0 +1,8 @@
+package duke.util;
+
+import duke.exception.DukeException;
+import duke.task.TaskList;
+
+public interface CheckedFunction {
+    String execute(TaskList t) throws DukeException;
+}
