+++ src/main/java/duke/util/Ui.java
@@ -1,3 +1,5 @@
+package duke.util;
+
 import java.util.Scanner;
 
 public class Ui {
